--- tasks/manifests/pixi-rope.json
@@ -1,4 +1,0 @@
-[
-    "src/pixi/extras/Strip.js",
-    "src/pixi/extras/Rope.js"
-]
