+++ tasks/options/concat.js
@@ -35,11 +35,6 @@ module.exports = {
         dest: '<%= modules_dir %>/pixi-main.js'
     },
     
-    pixiRope: {
-        src: require('../manifests/pixi-rope'),
-        dest: '<%= modules_dir %>/pixi-rope.js'
-    },
-
     pixiOutro: {
         src: require('../manifests/pixi-outro'),
         dest: '<%= modules_dir %>/pixi-outro.js'
